Fixed error in calculating comfort violation time for GausTrapReward
Restructured the way comfort violation such that it is not based on the
comfort penalty value, but instead directly based on if the temperatures of
the current step are outside the comfort range. This allows custom reward
signals that can take non-zero values while in the comfort range.

The comfort violation calculation is handled already in the reward calculating
function and is returned as part of the reward terms. In this commit all parts
of the sinergym codebase that needs to consider this change are updated accordingly